Move BookstorePublishHandler into its own publish module
- wanted to extract publishing from the handlers before beginning to
work on the cloning functionality which will live in its own module
